rye-devtools: move common functionality to a 'common' module
We are starting a common module that includes some functionality that
we will reuse when writing a script to fetch uv downloads.
