+++ rye-devtools/src/rye_devtools/find_downloads.py
@@ -10,16 +10,15 @@ import itertools
 import os
 import re
 import sys
-import time
 import unittest
 from dataclasses import dataclass
-from datetime import datetime, timezone
 from enum import StrEnum
-from typing import NamedTuple, Self
 from urllib.parse import unquote
 
 import httpx
 from httpx import HTTPStatusError
+
+from .common import Version, fetch, PlatformTriple
 
 
 def log(*args, **kwargs):
@@ -36,26 +35,6 @@ def batched(iterable, n):
         yield batch
 
 
-class PlatformTriple(NamedTuple):
-    arch: str
-    platform: str
-    environment: str | None
-    flavor: str | None
-
-
-class PythonVersion(NamedTuple):
-    major: int
-    minor: int
-    patch: int
-
-    @classmethod
-    def from_str(cls, version: str) -> Self:
-        return cls(*map(int, version.split(".", 3)))
-
-    def __neg__(self) -> Self:
-        return PythonVersion(-self.major, -self.minor, -self.patch)
-
-
 class PythonImplementation(StrEnum):
     CPYTHON = "cpython"
     PYPY = "pypy"
@@ -63,43 +42,12 @@ class PythonImplementation(StrEnum):
 
 @dataclass
 class PythonDownload:
-    version: PythonVersion
+    version: Version
     triple: PlatformTriple
     implementation: PythonImplementation
     filename: str
     url: str
     sha256: str | None = None
-
-
-async def fetch(client: httpx.AsyncClient, url: str) -> httpx.Response:
-    """Fetch from GitHub API with rate limit awareness."""
-    resp = await client.get(url, timeout=15)
-    if (
-        resp.status_code in [403, 429]
-        and resp.headers.get("x-ratelimit-remaining") == "0"
-    ):
-        # See https://docs.github.com/en/rest/using-the-rest-api/troubleshooting-the-rest-api?apiVersion=2022-11-28
-        if (retry_after := resp.headers.get("retry-after")) is not None:
-            log(f"Got retry-after header, retry in {retry_after} seconds.")
-            time.sleep(int(retry_after))
-
-            return await fetch(client, url)
-
-        if (retry_at := resp.headers.get("x-ratelimit-reset")) is not None:
-            utc = datetime.now(timezone.utc).timestamp()
-            retry_after = max(int(retry_at) - int(utc), 0)
-
-            log(f"Got x-ratelimit-reset header, retry in {retry_after} seconds.")
-            time.sleep(retry_after)
-
-            return await fetch(client, url)
-
-        log("Got rate limited but no information how long, wait for 2 minutes.")
-        time.sleep(60 * 2)
-        return await fetch(client, url)
-
-    resp.raise_for_status()
-    return resp
 
 
 class Finder:
@@ -182,7 +130,7 @@ class CPythonFinder(Finder):
 
     async def fetch_indygreg_downloads(self, pages: int = 100) -> list[PythonDownload]:
         """Fetch all the indygreg downloads from the release API."""
-        results: dict[PythonVersion, dict[tuple[str, str], list[PythonDownload]]] = {}
+        results: dict[Version, dict[tuple[str, str], list[PythonDownload]]] = {}
 
         for page in range(1, pages):
             log(f"Fetching indygreg release page {page}")
@@ -227,7 +175,7 @@ class CPythonFinder(Finder):
             return
 
         version_str, triple_str = match.groups()
-        version = PythonVersion.from_str(version_str)
+        version = Version.from_str(version_str)
         triple = cls.parse_triple(triple_str)
         if triple is None:
             return
@@ -375,7 +323,7 @@ class PyPyFinder(Finder):
         for version in versions:
             if not version["stable"]:
                 continue
-            python_version = PythonVersion.from_str(version["python_version"])
+            python_version = Version.from_str(version["python_version"])
             if python_version < (3, 7, 0):
                 continue
             for file in version["files"]:
@@ -418,7 +366,7 @@ class PyPyFinder(Finder):
 def render(downloads: list[PythonDownload]):
     """Render downloads.inc."""
 
-    def sort_key(download: PythonDownload) -> tuple[int, PythonVersion, PlatformTriple]:
+    def sort_key(download: PythonDownload) -> tuple[int, Version, PlatformTriple]:
         # Sort by implementation, version (latest first), and then by triple.
         impl_order = [PythonImplementation.PYPY, PythonImplementation.CPYTHON]
         return (
